+++ NEWS.md
@@ -1,7 +1,7 @@
 # Unreleased version
 
-- Data module: support reading many more file types beyond CSV
-- Data module: show user friendly error message when a data source fails to get read correctly 
+- Data module: support reading many more file types beyond CSV (.csv, .txt, .tsv, .sas7bdat, .xpt, .sav, .zsav, .dta, .por, .xls, .xlsx)
+- Data module: show user-friendly error message when a data source fails to get read correctly 
 - Data module: make preview and code always reflect the currently selected tab
 - Add all package dependencies that were previously missing